changed-enterpises
Changed the html to print a list of enterprises. In each row contains
the name of the enterprise and a text field to enter a password. If the
enterprise the user wants to access has a password it is checked against
what the user has entered and if they match the user sees the map,
otherwise the user is told that the passwords do not match. If the
enterprise does not have a password the user sees the map.